backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "argon2-cffi>=23.1.0",
    "asyncpg==0.29.0",
    "email-validator==2.1.1",
    "fastapi==0.103.1",
    "pydantic==2.7.0",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0",
    "redis==5.0.1",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]==2.0.23",
    "sse-starlette==1.6.5",
    "structlog>=24.4.0",
    "uvicorn==0.23.2",
]

[dependency-groups]
dev = [
    "httpx>=0.24,<0.25",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.12.11",
    "types-requests>=2.32.0.20250710",
    "types-pyyaml>=6.0.12.20250822",
    "types-python-jose>=3.4.0.20250626",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = [
    "B008", # FastAPI Depends(...) is commonly used in parameter defaults
    "E501", # Handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
